Add specifiable complete arguments at global activation/registration
Makes it possible to remove defaults using global activation and no just
with registration. Also enables generically customizing complete arguments.
Feedback is given when non-default arguments are specified when globally
activating.

The --no-defaults option is kept for backward compatibility and ease of
use.

Changes are best seen running activate-global-python-argcomplete --dest=-
with various options and looking at the last line in the ouput, the call
to complete. e.g:
* activate-global-python-argcomplete --dest=-
  complete -o nospace -o default -o bashdefault -D -F _python_argcomplete_global
* activate-global-python-argcomplete --dest=- --no-defaults
  complete -o nospace -o bashdefault -D -F _python_argcomplete_global
* activate-global-python-argcomplete --dest=- --complete-arguments --o nospace
  complete -o nospace -D -F _python_argcomplete_global
* activate-global-python-argcomplete --dest=- --complete-arguments
  complete -D -F _python_argcomplete_global

In summary, the call to complete in python-argcomplete.sh is removed, and
appended to the file with appropriate arguments after the file is copied.
The same options are added to register-python-argcomplete for consistency.

Note: the difference remains that by default the register script passes
neither -D nor -o bashdefault to complete, whereas the activate script does
pass them.