Add debug_stream argument
Allow debug_stream to be optionally set in CompletionFinder.__call__.
By default, this is still set to os.fdopen(9), but now it can be
overridden to a different i/o stream instead. May be useful for
unit tests, especially as os.fdopen(9) is likely to conflict
with pytest plugins and lead to obscure errors.

Closes #393